Update 4400Proj.py
The Main page is mostly done I just need to add the project/course part at the bottom which will mostly be sql stuff. I also added that once you hit login button the login window is closed and the main page window is opened (same for register). and the two new functions apply and reset are called by the main page window