Update 4400Proj.py
Updated dept function and Edit_profile function to allow for the automatic update of the Department name on the Edit profile page.  This edit also includes the SQL to update the Year and Major of a student in the database if they change that info on the Edit profile page.